requirements.txt
absl-py==1.4.0
mdurl==0.1.2
ortools==9.1.9490
pika==1.3.2
protobuf==3.20.1
Pygments==2.15.1
python-dotenv==1.0.0
rich==13.4.1
